+++ vehicle-python/src/vehicle_lang/__init__.py
@@ -1,7 +1,39 @@
 from typing import List
 
 from ._version import VERSION as VERSION
+from .compile import load_loss_function as load_loss_function
+from .error import BuiltinUnsupported as BuiltinUnsupported
+from .error import InternalError as InternalError
+from .error import PropertyNameNotFound as PropertyNameNotFound
+from .error import SessionClosed as SessionClosed
+from .error import SessionUsed as SessionUsed
+from .error import VehicleError as VehicleError
+from .typing import AnySamplers as AnySamplers
+from .typing import DeclarationName as DeclarationName
+from .typing import DifferentiableLogic as DifferentiableLogic
+from .typing import QuantifiedVariableName as QuantifiedVariableName
+from .typing import Sampler as Sampler
+from .typing import Verifier as Verifier
+from .verify import verify as verify
 
 __all__: List[str] = [
     "VERSION",
+    # Compile
+    "load_loss_function",
+    # Verify
+    "verify",
+    # Error types
+    "VehicleError",
+    "SessionClosed",
+    "SessionUsed",
+    "BuiltinUnsupported",
+    "InternalError",
+    "PropertyNameNotFound",
+    # Argument types
+    "DeclarationName",
+    "QuantifiedVariableName",
+    "Sampler",
+    "AnySamplers",
+    "DifferentiableLogic",
+    "Verifier",
 ]