+++ vehicle-python/tests/test_lossdl2_exec_tf_bounded.py
@@ -2,7 +2,7 @@ import random
 from pathlib import Path
 from typing import Any, Dict, Iterator
 
-from vehicle_lang.compile import Target, to_python
+import vehicle_lang as vcl
 
 
 def test_lossdl2_exec_tf_bounded() -> None:
@@ -29,12 +29,12 @@ def test_lossdl2_exec_tf_bounded() -> None:
             for _ in range(0, 10):
                 yield random.uniform(0.5, 0.5)
 
-        bounded = to_python(
-            specification_path, target=Target.LOSS_DL2, samplers={"x": sampler_for_x}
-        )["bounded"]
-
-        def bounded_loss() -> Any:
-            return bounded(f)
+        bounded_loss = vcl.load_loss_function(
+            specification_path,
+            property_name="bounded",
+            target=vcl.DifferentiableLogic.DL2,
+            samplers={"x": sampler_for_x},
+        )
 
         # Prepare training data
         batch_size = 1
@@ -75,7 +75,7 @@ def test_lossdl2_exec_tf_bounded() -> None:
                         x_batch_train, training=True
                     )  # Outputs for this minibatch
                     ce_loss_value = ce_batch_loss(y_batch_train, outputs)
-                    bounded_value = bounded_loss()
+                    bounded_value = bounded_loss(network=f)
                     weighted_loss = (
                         ce_loss_value * ce_loss_weight + bounded_value * bounded_weight
                     )