+++ vehicle-python/tests/test_lossdl2_exec_tf_mnist_robustness.py
@@ -1,46 +1,15 @@
 from pathlib import Path
 from typing import Any, Dict, Iterator, Tuple, cast
 
+import vehicle_lang as vcl
 from typing_extensions import TypeAlias
-from vehicle_lang.compile import Target, to_python
 
 GOLDEN_PATH = (
     Path(__file__).parent.parent / "vendor" / "vehicle" / "tests" / "golden" / "compile"
 )
 MNIST_ROBUSTNESS = GOLDEN_PATH / "mnist-robustness" / "spec.vcl"
 
-# fmt: off
-Image: TypeAlias = Tuple[
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-    Tuple[float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float,float],
-]
-# fmt: on
+Image: TypeAlias = Tuple[Tuple[float, ...], ...]
 
 Perturbation: TypeAlias = Image
 
@@ -99,23 +68,18 @@ def test_lossdl2_exec_tf_mnist_robustness() -> None:
     def classifier(image: Image) -> LabelDistribution:
         return one_hot(0)
 
-    # def robust(
-    #   n: int,
-    #   classifier: Callable[[Image], Label],
-    #   epsilon: float,
-    #   trainingImages: Sequence[Image],
-    #   trainingLabels: Sequence[Label]
-    # ):
-    robust = to_python(
+    robust_loss = vcl.load_loss_function(
         MNIST_ROBUSTNESS,
-        target=Target.LOSS_DL2,
+        property_name="robust",
+        target=vcl.DifferentiableLogic.DL2,
         samplers={"pertubation": sampler_for_pertubation},
-    )["robust"]
-    robust_loss = robust(
+    )
+
+    loss = robust_loss(
         n=1,
         classifier=classifier,
         epsilon=0.001,
         trainingImages=(ZEROES_28X28,),
         trainingLabels=(0,),
     )
-    print(robust_loss)
+    print(loss)