+++ vehicle-python/tests/test_lossdl2_load.py
@@ -2,6 +2,8 @@ import os
 from pathlib import Path
 
 import pytest
+import vehicle_lang as vcl
+import vehicle_lang.ast as vcla
 
 GOLDEN_PATH = (
     Path(__file__).parent.parent / "vendor" / "vehicle" / "tests" / "golden" / "compile"
@@ -16,6 +18,7 @@ GOLDEN_LOSS_FUNCTION_FILES = GOLDEN_PATH.glob(os.path.join("**", "DL2Loss.vcl.go
 def test_interpret_load(dl2loss_specification_path: Path) -> None:
     specification_path = dl2loss_specification_path.parent / "spec.vcl"
     print(f"Load {specification_path}")
-    from vehicle_lang.session import Target, load
-
-    load(specification_path, target=Target.LOSS_DL2)
+    vcla.load(
+        specification_path,
+        target=vcl.DifferentiableLogic.DL2,
+    )